refactor: 어댑터 파일 생성
